widgets: Prevent UB in send event due to type promotion
Clang reports that passing an argument subject to default type
promotion to va_start results in undefined behavior. Because sub_type
is uint16_t and the function is variadic, it will be promoted to
unsigned int.

This can be avoided by simply declaring the parameter as a full width
integer.

+++ include/widgets/gp_widget_event.h
@@ -145,7 +145,7 @@ static inline int gp_widget_send_event(gp_widget *self,
  * @return The return value from application event handler.
  */
 static inline int gp_widget_send_widget_event(gp_widget *self,
-				              uint16_t sub_type,
+				              unsigned int sub_type,
 				              ...)
 {
 	if (!self->on_event)
